+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/SecurityExtensionTest.php
@@ -123,35 +123,6 @@ class SecurityExtensionTest extends TestCase
         $this->assertFalse($container->hasDefinition('security.access.role_hierarchy_voter'));
     }
 
-    /**
-     * @group legacy
-     * @expectedDeprecation Setting logout_on_user_change to false is deprecated as of 3.4 and will always be true in 4.0. Set logout_on_user_change to true in your firewall configuration.
-     */
-    public function testDeprecationForUserLogout()
-    {
-        $container = $this->getRawContainer();
-
-        $container->loadFromExtension('security', array(
-            'providers' => array(
-                'default' => array('id' => 'foo'),
-            ),
-
-            'firewalls' => array(
-                'some_firewall' => array(
-                    'pattern' => '/.*',
-                    'http_basic' => null,
-                    'logout_on_user_change' => false,
-                ),
-            ),
-        ));
-
-        $container->compile();
-    }
-
-    /**
-     * @group legacy
-     * @expectedDeprecation Firewall "some_firewall" is configured as "stateless" but the "switch_user.stateless" key is set to false. Both should have the same value, the firewall's "stateless" value will be used as default value for the "switch_user.stateless" key in 4.0.
-     */
     public function testSwitchUserNotStatelessOnStatelessFirewall()
     {
         $container = $this->getRawContainer();
@@ -172,31 +143,8 @@ class SecurityExtensionTest extends TestCase
         ));
 
         $container->compile();
-    }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     * @expectedExceptionMessage Not configuring explicitly the provider on "default" firewall is ambiguous as there is more than one registered provider.
-     */
-    public function testDeprecationForAmbiguousProvider()
-    {
-        $container = $this->getRawContainer();
-
-        $container->loadFromExtension('security', array(
-            'providers' => array(
-                'first' => array('id' => 'foo'),
-                'second' => array('id' => 'bar'),
-            ),
-
-            'firewalls' => array(
-                'default' => array(
-                    'http_basic' => null,
-                    'logout_on_user_change' => true,
-                ),
-            ),
-        ));
-
-        $container->compile();
+        $this->assertTrue($container->getDefinition('security.authentication.switchuser_listener.some_firewall')->getArgument(9));
     }
 
     protected function getRawContainer()