+++ src/Symfony/Component/Filesystem/Tests/FilesystemTest.php
@@ -1141,33 +1141,21 @@ class FilesystemTest extends FilesystemTestCase
     }
 
     /**
-     * @group legacy
-     * @dataProvider provideLegacyPathsForMakePathRelativeWithRelativePaths
-     * @expectedDeprecation Support for passing relative paths to Symfony\Component\Filesystem\Filesystem::makePathRelative() is deprecated since version 3.4 and will be removed in 4.0.
+     * @expectedException \Symfony\Component\Filesystem\Exception\InvalidArgumentException
+     * @expectedExceptionMessage The start path "var/lib/symfony/src/Symfony/Component" is not absolute.
      */
-    public function testMakePathRelativeWithRelativePaths($endPath, $startPath, $expectedPath)
+    public function testMakePathRelativeWithRelativeStartPath()
     {
-        $path = $this->filesystem->makePathRelative($endPath, $startPath);
-
-        $this->assertEquals($expectedPath, $path);
+        $this->assertSame('../../../', $this->filesystem->makePathRelative('/var/lib/symfony/', 'var/lib/symfony/src/Symfony/Component'));
     }
 
-    public function provideLegacyPathsForMakePathRelativeWithRelativePaths()
+    /**
+     * @expectedException \Symfony\Component\Filesystem\Exception\InvalidArgumentException
+     * @expectedExceptionMessage The end path "var/lib/symfony/" is not absolute.
+     */
+    public function testMakePathRelativeWithRelativeEndPath()
     {
-        return array(
-            array('usr/lib/symfony/', 'var/lib/symfony/src/Symfony/Component', '../../../../../../usr/lib/symfony/'),
-            array('aa/bb', 'aa/cc', '../bb/'),
-            array('aa/cc', 'bb/cc', '../../aa/cc/'),
-            array('aa/bb', 'aa/./cc', '../bb/'),
-            array('aa/./bb', 'aa/cc', '../bb/'),
-            array('aa/./bb', 'aa/./cc', '../bb/'),
-            array('../../', '../../', './'),
-            array('../aa/bb/', 'aa/bb/', '../../../aa/bb/'),
-            array('../../../', '../../', '../'),
-            array('', '', './'),
-            array('', 'aa/', '../'),
-            array('aa/', '', 'aa/'),
-        );
+        $this->assertSame('../../../', $this->filesystem->makePathRelative('var/lib/symfony/', '/var/lib/symfony/src/Symfony/Component'));
     }
 
     public function testMirrorCopiesFilesAndDirectoriesRecursively()