+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -14,6 +14,9 @@ CHANGELOG
  * removed command `init:acl` along with `InitAclCommand` class
  * removed `acl` configuration key and related services, use symfony/acl-bundle instead
  * removed auto picking the first registered provider when no configured provider on a firewall and ambiguous
+ * the firewall option `logout_on_user_change` is now always true, which will trigger a logout if the user changes
+   between requests
+ * the `switch_user.stateless` firewall option is `true` for stateless firewalls
 
 3.4.0
 -----